requirements.txt
# Review Phim Pro — Dependencies
# Core
PyQt6>=6.6.0
PyQt6-Qt6>=6.6.0
PyQt6-sip>=13.6.0
PyQt6-WebEngine>=6.6.0

# Translation
deep-translator>=1.11.0
google-generativeai>=0.5.0
requests>=2.31.0

# Video downloader (Douyin / TikTok)
yt-dlp>=2024.7.1

# Voice
edge-tts>=6.1.0
gTTS>=2.5.0
pydub>=0.25.1
elevenlabs>=1.0.0

# Video
opencv-python>=4.9.0
moviepy>=1.0.3

# Subtitle Extraction
openai-whisper>=20231117

# OCR (optional — requires paddlepaddle)
# paddleocr>=2.7.0
# paddlepaddle>=2.6.0

# AI
torch>=2.2.0
